@@ -12,7 +12,7 @@ dependencies = [
     "httpx>=0.27.0",
     "rich>=13.7.1",
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 readme = "README.md"
 license = {text = "MIT"}
 
